refactor tests (#395)
