+++ frontend/test/integration/createCommunicationTemplate.test.tsx
@@ -1,11 +1,11 @@
 import "@testing-library/jest-dom";
 import { render, waitFor } from "@testing-library/react";
-import CommunicationTemplateCreate from "../src/pages/communicationTemplate/create";
+import CommunicationTemplateCreate from "../../src/pages/communicationTemplate/create";
 import userEvent from "@testing-library/user-event";
 import mockRouter from "next-router-mock";
-import { CommunicationTemplateEntity } from "../src/api/entities/CommunicationTemplateEntity";
+import { CommunicationTemplateEntity } from "../../src/api/entities/CommunicationTemplateEntity";
 import mockAxios from "jest-mock-axios";
-import apiPaths from "../src/properties/apiPaths";
+import apiPaths from "../../src/properties/apiPaths";
 
 jest.mock("next/router", () => require("next-router-mock"));
 
@@ -21,7 +21,7 @@ describe("create communication template", () => {
 
     it("Sends the form and calls post", async () => {
         const spy = jest.spyOn(
-            require("../src/handlers/createCommunicationTemplateSubmitHandler"),
+            require("../../src/handlers/createCommunicationTemplateSubmitHandler"),
             "createCommunicationTemplateSubmitHandler"
         );
