+++ frontend/test/integration/home.test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react";
-import Home from "../src/pages/home";
+import Home from "../../src/pages/home";
 import "@testing-library/jest-dom";
 
 describe("Home page", () => {
