+++ frontend/test/integration/projects.test.tsx
@@ -1,8 +1,8 @@
 import "@testing-library/jest-dom";
 import mockAxios from "jest-mock-axios";
 import { render, screen } from "@testing-library/react";
-import ProjectPage from "../src/pages/projects";
-import ProjectIDPage from "../src/pages/projects/[id]";
+import ProjectPage from "../../src/pages/projects";
+import ProjectIDPage from "../../src/pages/projects/[id]";
 
 jest.mock("next/router", () => require("next-router-mock"));
