+++ frontend/test/integration/studentFilterComponent.test.tsx
@@ -1,6 +1,6 @@
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
-import { StudentFilterComponent } from "../src/components/studentFilterComponent";
+import { StudentFilterComponent } from "../../src/components/studentFilterComponent";
 
 jest.mock("next/router", () => require("next-router-mock"));
