+++ frontend/test/integration/usersList.test.tsx
@@ -3,9 +3,9 @@ import userEvent from "@testing-library/user-event";
 import mockAxios from "jest-mock-axios";
 import { act, render, screen, waitFor } from "@testing-library/react";
 import { makeCacheFree } from "./Provide";
-import UsersOverview from "../src/components/usersOverview";
-import UserComponent from "../src/components/manageUserComponent";
-import { IUser, UserRole } from "../src/api/entities/UserEntity";
+import UsersOverview from "../../src/components/usersOverview";
+import UserComponent from "../../src/components/manageUserComponent";
+import { IUser, UserRole } from "../../src/api/entities/UserEntity";
 import {
     getBaseBadRequestResponse,
     getBaseNoContentResponse,
@@ -13,7 +13,6 @@ import {
     getBaseUser,
 } from "./TestEntityProvider";
 import { AxiosResponse } from "axios";
-import { ReasonPhrases, StatusCodes } from "http-status-codes";
 
 jest.mock("next/router", () => require("next-router-mock"));
